Move synchronized timer and controller to components/core
