+++ src/components/core/synchronized-timer.ts
