+++ src/components/lineup-game-live.ts
@@ -46,6 +46,8 @@ import {
   toggleClock,
 } from '../slices/live/index.js';
 import { ConnectStoreMixin } from './core/connect-mixin.js';
+import { synchronizedTimerContext, SynchronizedTimerNotifier } from './core/synchronized-timer.js';
+import { SynchronizedTriggerController } from './core/timer-controller.js';
 import './lineup-game-clock.js';
 import { ClockEndPeriodEvent, ClockPeriodData } from './lineup-game-clock.js';
 import './lineup-game-events.js';
@@ -57,8 +59,6 @@ import './lineup-player-list.js';
 import { LineupPlayerList } from './lineup-player-list.js';
 import { playerResolverContext } from './player-resolver.js';
 import { SharedStyles } from './shared-styles.js';
-import { synchronizedTimerContext, SynchronizedTimerNotifier } from './synchronized-timer.js';
-import { SynchronizedTriggerController } from './timer-controller.js';
 
 // This element is connected to the Redux store.
 @customElement('lineup-game-live')
